tests/wpt/mozilla/tests/bluetooth/getPrimaryService/disconnect-called-before.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => {
        gattServer.disconnect();
        return promise_rejects_dom(t, 'NetworkError', gattServer.getPrimaryService(heart_rate.name));
    });
}, 'disconnect() called before getPrimaryService. Reject with NetworkError.');
</script>
